+++ pablo.h
@@ -41,7 +41,7 @@ extern int pablo_rid;   // active row id
 
 // histogram with agrupation rates a given int repeats
 extern long unsigned pablo_grouping_hist[PABLO_MAX_GROUPING];
-extern long unsigned pablo_occurrences;
+extern int pablo_occurrences;
 extern long unsigned int pablo_unocurrences_grouping_hist[PABLO_MAX_GROUPING];
 extern int pablo_unoccurrences;
